requirements.txt
attrs==22.1.0
aws-cdk-lib==2.28.1
cattrs==22.2.0
constructs==10.1.176
exceptiongroup==1.0.4
jsii==1.71.0
publication==0.0.3
python-dateutil==2.8.2
python-dotenv==0.21.0
six==1.16.0
typeguard==2.13.3
typing_extensions==4.4.0
